backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pydantic==2.5.2
openai==1.3.7
python-dotenv==1.0.0
httpx==0.25.1
beautifulsoup4==4.12.2
lxml==4.9.3
